requirements.txt
streamlit>=1.32.0
plotly>=5.18.0
pandas>=2.1.0
chromadb>=0.4.22
groq>=0.4.0
python-dotenv>=1.0.0
